More rename and fix tests on headless runs
